+++ Aufgabe 11/Data.py
